clang: only implicitly include the selected GCC multilib crtbegin.o
Previously we would include up to three common crtbegin.o files that exist
in the right location, and hope that one of them was correct.

Since Clang tells us which multilib it selected, just upload that one.
This works in more cases than the old hack.

Also, if crtbegin.o is not found, then return an error (don't just log it).

Change-Id: I096a362ddddc898109e8942a96e8e74ee87ef2a5